Collapse symbol layer into media layer
I don't do data entry, so most of the symbol layer is useless for me.
The only times that I ever enter it are to use the function keys, which
are few. I also accidentally enable symbol layer all the time, so it
would be better if I could use a different key for its functionality.